+++ addon/initializers/local-store.js
@@ -1,0 +1,12 @@
+import LocalStore from '../stores/local-store';
+
+export function initialize(application) {
+  //Register factory for local Store.
+  application.register('store:local', LocalStore);
+}
+
+export default {
+  name: 'local-store',
+  before: 'syncer',
+  initialize
+};