+++ tests/unit/utils/generate-unique-id-test.js
@@ -1,0 +1,10 @@
+import generateUniqueId from 'ember-flexberry-data/utils/generate-unique-id';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | generate unique id');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = generateUniqueId();
+  assert.ok(result);
+});